+++ finality-provider/service/fp_instance.go
@@ -11,6 +11,7 @@ import (
 	"github.com/avast/retry-go/v4"
 	bbntypes "github.com/babylonlabs-io/babylon/types"
 	ftypes "github.com/babylonlabs-io/babylon/x/finality/types"
+	fppath "github.com/babylonlabs-io/finality-provider/lib/math"
 	"github.com/btcsuite/btcd/btcec/v2"
 	"github.com/gogo/protobuf/jsonpb"
 	"go.uber.org/atomic"
@@ -202,6 +203,18 @@ func (fp *FinalityProviderInstance) finalitySigSubmissionLoop() {
 			if fp.hasProcessed(b) {
 				continue
 			}
+
+			activationBlkHeight, err := fp.cc.QueryFinalityActivationBlockHeight()
+			if err != nil {
+				fp.reportCriticalErr(fmt.Errorf("failed to get activation height during fast sync %w", err))
+				continue
+			}
+
+			// check if it is allowed to send finality
+			if b.Height < activationBlkHeight {
+				continue
+			}
+
 			// check whether the finality provider has voting power
 			hasVp, err := fp.hasVotingPower(b)
 			if err != nil {
@@ -374,16 +387,9 @@ func (fp *FinalityProviderInstance) tryFastSync(targetBlock *types.BlockInfo) (*
 		return nil, nil
 	}
 
-	lastFinalizedHeight := lastFinalizedBlocks[0].Height
-	lastProcessedHeight := fp.GetLastProcessedHeight()
-
-	// get the startHeight from the maximum of the lastVotedHeight and
-	// the lastFinalizedHeight plus 1
-	var startHeight uint64
-	if lastFinalizedHeight < lastProcessedHeight {
-		startHeight = lastProcessedHeight + 1
-	} else {
-		startHeight = lastFinalizedHeight + 1
+	startHeight, err := fp.fastSyncStartHeight(lastFinalizedBlocks[0].Height)
+	if err != nil {
+		return nil, err
 	}
 
 	if startHeight > targetBlock.Height {
@@ -393,6 +399,18 @@ func (fp *FinalityProviderInstance) tryFastSync(targetBlock *types.BlockInfo) (*
 	fp.logger.Debug("the finality-provider is entering fast sync")
 
 	return fp.FastSync(startHeight, targetBlock.Height)
+}
+
+func (fp *FinalityProviderInstance) fastSyncStartHeight(lastFinalizedHeight uint64) (uint64, error) {
+	lastProcessedHeight := fp.GetLastProcessedHeight()
+
+	finalityActivationBlkHeight, err := fp.cc.QueryFinalityActivationBlockHeight()
+	if err != nil {
+		return 0, err
+	}
+
+	// return the max start height by checking the finality activation block height
+	return fppath.MaxUint64(lastProcessedHeight+1, lastFinalizedHeight+1, finalityActivationBlkHeight), nil
 }
 
 func (fp *FinalityProviderInstance) hasProcessed(b *types.BlockInfo) bool {
@@ -597,11 +615,16 @@ func (fp *FinalityProviderInstance) CommitPubRand(tipHeight uint64) (*types.TxRe
 		return nil, nil
 	}
 
+	activationBlkHeight, err := fp.cc.QueryFinalityActivationBlockHeight()
+	if err != nil {
+		return nil, err
+	}
+
 	// generate a list of Schnorr randomness pairs
 	// NOTE: currently, calling this will create and save a list of randomness
 	// in case of failure, randomness that has been created will be overwritten
 	// for safety reason as the same randomness must not be used twice
-	pubRandList, err := fp.getPubRandList(startHeight, fp.cfg.NumPubRand)
+	pubRandList, err := fp.getPubRandList(fppath.MaxUint64(startHeight, activationBlkHeight), fp.cfg.NumPubRand)
 	if err != nil {
 		return nil, fmt.Errorf("failed to generate randomness: %w", err)
 	}