+++ CHANGELOG.md
@@ -37,13 +37,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
-### Misc Improvements
+### Improvements
 
 * [#104](https://github.com/babylonlabs-io/finality-provider/pull/104) Print fpd binary version
 
+
+## v0.9.0
+
+### Improvements
+
+* [#101](https://github.com/babylonlabs-io/finality-provider/pull/101) Add finality activation
+height check in finality voting and commit pub rand start height and bump Babylon version to
+v0.14.0
+
 ## v0.8.0
 
-### Misc Improvements
+### Improvements
 
 * [#97](https://github.com/babylonlabs-io/finality-provider/pull/97) Bump Babylon version to v0.13.0
 * [#90](https://github.com/babylonlabs-io/finality-provider/pull/90) CLI edit finality provider